Move trainable mask.
